+++ tests/unit/models/generation/test_megatron_generation.py
@@ -14,6 +14,7 @@
 
 import gc
 from copy import deepcopy
+from types import SimpleNamespace
 
 import pytest
 import torch
@@ -22,7 +23,7 @@ from nemo_rl.algorithms.grpo import refit_policy_generation
 from nemo_rl.algorithms.utils import get_tokenizer
 from nemo_rl.distributed.batched_data_dict import BatchedDataDict
 from nemo_rl.distributed.virtual_cluster import RayVirtualCluster
-from nemo_rl.models.generation.megatron import MegatronGeneration
+from nemo_rl.models.generation.megatron import MegatronGeneration, megatron_generation
 from nemo_rl.models.generation.megatron.config import (
     dedicated_inference_megatron_cfg,
 )
@@ -630,3 +631,112 @@ def test_megatron_generation_non_colocated_refit(
             print(f"Error during generation_cluster shutdown: {e}")
         gc.collect()
         torch.cuda.empty_cache()
+
+
+class _CapturingPortHolder:
+    """Stand-in for the RemoteHeldPortReservation actor.
+
+    Records the scheduling strategy it is pinned to and returns a fixed
+    (ip, port) instead of binding a real socket on a real placement group.
+    """
+
+    last_scheduling_strategy = None
+
+    @classmethod
+    def options(cls, *, scheduling_strategy):
+        cls.last_scheduling_strategy = scheduling_strategy
+        return cls
+
+    @classmethod
+    def remote(cls):
+        return SimpleNamespace(
+            address=SimpleNamespace(remote=lambda: ("10.0.0.5", 4321))
+        )
+
+
+def _rank0_bundle_via_worker_group(sorted_bundle_indices, group_size):
+    """The bundle RANK 0 actually lands on, reconstructed from the live code.
+
+    Mirrors lm_policy.py's tied_groups[0] for a unified PG and RayWorkerGroup's
+    default first-worker tuple otherwise -- the two branches
+    reserve_http_server_address must agree with.
+
+    Deliberately a hand-copy rather than a call into the code under test (or a
+    shared helper): sharing the implementation would make the assertion a
+    tautology. The cost is that this copy is frozen -- if the worker-group
+    placement rule ever changes, update it by hand; grpo.py's runtime
+    served-vs-reserved URL check is the net for that direction.
+    """
+    if sorted_bundle_indices is not None:
+        # lm_policy.py: tied_groups = [(i // group_size, [b]) for i, b in ...]
+        tied_groups = [
+            (i // group_size, [bundle_idx])
+            for i, bundle_idx in enumerate(sorted_bundle_indices)
+        ]
+    else:
+        # RayWorkerGroup.__init__: bundle_indices_list.append((i, [bundle_idx]))
+        # with i and bundle_idx both starting at 0.
+        tied_groups = [(0, [0])]
+    pg_idx, local_bundle_indices = tied_groups[0]
+    return pg_idx, local_bundle_indices[0]
+
+
+@pytest.fixture
+def patched_holder(monkeypatch):
+    monkeypatch.setattr(
+        megatron_generation, "RemoteHeldPortReservation", _CapturingPortHolder
+    )
+    # ray.get here only unwraps the holder's (ip, port); no real Ray involved.
+    monkeypatch.setattr(megatron_generation.ray, "get", lambda ref: ref)
+    _CapturingPortHolder.last_scheduling_strategy = None
+    return _CapturingPortHolder
+
+
+@pytest.mark.parametrize(
+    "sorted_bundle_indices",
+    [
+        # Unified cross-node PG: topology sort can make rank 0 a bundle other
+        # than 0, so a naive "bundle 0" prediction would bind the wrong node.
+        [3, 1, 0, 2],
+        # Per-node PGs: no sorted indices, rank 0 is bundle 0 of the first PG.
+        None,
+    ],
+)
+def test_reserve_http_server_address_pins_rank0_bundle(
+    patched_holder, sorted_bundle_indices
+):
+    """reserve_http_server_address's rank-0 prediction must match real placement.
+
+    reserve_http_server_address publishes the OpenAI server URL to NeMo Gym
+    *before* any worker exists, pinning a port holder to the (placement_group,
+    bundle) it predicts rank 0 will occupy. That prediction is a second,
+    hand-written copy of the rank-0 placement lm_policy.py / RayWorkerGroup
+    actually perform: if the two ever disagree the holder binds the wrong node,
+    the pre-published URL is unreachable, and grpo.py fails loud at runtime.
+    Pins the prediction to a hand-reconstruction of that placement so the two
+    copies cannot silently drift -- no GPU or mcore extra needed
+    (MegatronGeneration imports without megatron.core).
+    """
+    placement_groups = ["PG0", "PG1"]
+    cluster = SimpleNamespace(
+        num_gpus_per_node=8,
+        _sorted_bundle_indices=sorted_bundle_indices,
+        get_placement_groups=lambda: placement_groups,
+    )
+    config = {"generation": {"colocated": {"enabled": True}}}
+
+    url, port, holder = MegatronGeneration.reserve_http_server_address(cluster, config)
+
+    expected_pg_idx, expected_bundle_index = _rank0_bundle_via_worker_group(
+        sorted_bundle_indices, cluster.num_gpus_per_node
+    )
+    strategy = patched_holder.last_scheduling_strategy
+    # The holder -- and thus the pre-published URL's node -- must sit on the
+    # exact (placement_group, bundle) rank 0 will occupy.
+    assert expected_pg_idx == 0  # rank 0 is always in the first placement group
+    assert strategy.placement_group is placement_groups[expected_pg_idx]
+    assert strategy.placement_group_bundle_index == expected_bundle_index
+
+    assert url == "http://10.0.0.5:4321/v1"
+    assert port == 4321
+    assert holder.address.remote() == ("10.0.0.5", 4321)
